# backend/swagger-export.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/me": {
      "get": {
        "operationId": "UsersController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current authenticated user profile",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_updateMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update your own user profile or switch workspace",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/me/settings": {
      "get": {
        "operationId": "UsersController_getMySettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user settings",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_updateMySettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update current user settings",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/licenses/available": {
      "get": {
        "operationId": "UsersController_getAvailableLicenses",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List available (unassigned) licenses",
        "tags": [
          "Users"
        ]
      }
    },
    "/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "root",
                "billing_admin",
                "admin",
                "manager",
                "caller",
                "marketing"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "Working",
                "On Leave",
                "Invited",
                "Deleted"
              ]
            }
          },
          {
            "name": "licenseType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "Whatsapp Chat Sync",
                "Core crm"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all team members",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a user by ID",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a user profile or role (admin only)",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove (soft-delete) a user (admin only)",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}/activity": {
      "get": {
        "operationId": "UsersController_getUserActivity",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get activity logs for a specific user",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/invite": {
      "post": {
        "operationId": "UsersController_invite",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Invite a new user to the organization (admin only)",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/licenses/assign": {
      "post": {
        "operationId": "UsersController_assignLicense",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignLicenseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assign a license to a user (admin only)",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/invitations": {
      "post": {
        "operationId": "InvitationsController_createInvitation",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new invitation",
        "tags": [
          "invitations"
        ]
      }
    },
    "/invitations/pending": {
      "get": {
        "operationId": "InvitationsController_listPending",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List pending invitations for the organization",
        "tags": [
          "invitations"
        ]
      }
    },
    "/invitations/{token}": {
      "get": {
        "operationId": "InvitationsController_verifyToken",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Verify an invitation token",
        "tags": [
          "invitations"
        ]
      }
    },
    "/invitations/{token}/accept": {
      "post": {
        "operationId": "InvitationsController_acceptInvitation",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Accept an invitation",
        "tags": [
          "invitations"
        ]
      }
    },
    "/invitations/{id}": {
      "delete": {
        "operationId": "InvitationsController_cancelInvitation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel/Delete an invitation",
        "tags": [
          "invitations"
        ]
      }
    },
...
  },
  "info": {
    "title": "Telecalling CRM API",
    "description": "Scalable REST API for the Telecalling CRM platform",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "NotificationPreferencesDto": {
        "type": "object",
        "properties": {
          "paymentPending": {
            "type": "boolean",
            "example": true
          },
          "paymentCompleted": {
            "type": "boolean",
            "example": true
          },
          "paymentFailed": {
            "type": "boolean",
            "example": true
          },
          "newLead": {
            "type": "boolean",
            "example": true
          },
          "callReminder": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "UserSettingsDto": {
        "type": "object",
        "properties": {
          "emailHandler": {
            "type": "string",
            "enum": [
              "web",
              "mobile"
            ],
            "example": "web"
          },
          "whatsappHandler": {
            "type": "string",
            "enum": [
              "web",
              "mobile"
            ],
            "example": "mobile"
          },
          "notifications": {
            "$ref": "#/components/schemas/NotificationPreferencesDto"
          }
        }
      },
      "InviteUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "ali@company.com"
          },
          "name": {
            "type": "string",
            "example": "Ali Hassan"
          },
          "initials": {
            "type": "string",
            "example": "AH"
          },
          "password": {
            "type": "string",
            "example": "Password123!"
          },
          "phone": {
            "type": "string",
            "example": "+923001234567"
          },
          "role": {
            "type": "string",
            "enum": [
              "root",
              "billing_admin",
              "admin",
              "manager",
              "caller",
              "marketing"
            ],
            "default": "caller"
          },
          "permissionTemplateId": {
            "type": "string",
            "description": "Permission template ID to assign"
          },
          "licenseType": {
            "type": "string",
            "enum": [
              "Whatsapp Chat Sync",
              "Core crm"
            ]
          }
        },
        "required": [
          "email"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Ali Hassan"
          },
          "role": {
            "type": "string",
            "enum": [
              "root",
              "billing_admin",
              "admin",
              "manager",
              "caller",
              "marketing"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "Working",
              "On Leave",
              "Invited",
              "Deleted"
            ]
          },
          "permissionTemplateId": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        }
      },
      "AssignLicenseDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "User ID to assign license to"
          },
          "licenseType": {
            "type": "string",
            "enum": [
              "Whatsapp Chat Sync",
              "Core crm"
            ]
          },
          "expiryDate": {
            "type": "string",
            "example": "2027-03-17"
          }
        },
        "required": [
          "userId",
          "licenseType",
          "expiryDate"
        ]
      },
...
    }
  }
}